fix up inference tests for the move
