+++ holopy/fitting/__init__.py
@@ -38,4 +38,3 @@ from model import Model, Parametrization
 from parameter import Parameter, par, ComplexParameter
 from fit_series import fit_series
 from minimizer import Nmpfit
-import prior
